du: support blocksize env variable
help states that blocksize can be used. This commit adds the
implementation so it can be.

Tests require setting of env variables. I dislike how this is done due
to the fact they may not be cleared properly and it will occasionally
break when run in parallel but this is how tests/test_dircolors.rs does
it